backend/requirements.txt
Flask
flask-cors
flask-limiter
requests
tldextract
python-dotenv
python-whois
